# docs/openapi.yaml
openapi: 3.0.3
info:
  title: LCDP Restaurant API
  version: 1.0.0
  description: REST API for restaurant operations (users, tables, menu, inventory, orders, billing).
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
...
paths:
...
  /v1/orders/{id}:
    get:
      security:
        - sessionCookie: []
      summary: Get order details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Success
    delete:
      security:
        - sessionCookie: []
      summary: Delete order
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Success
...